# Speed up brew --prefix