@MikeMcQuaid 2017-07-25
Thanks for your first contribution to Homebrew/brew, @kevinburke! Without people like you submitting PRs we couldn't run this project. You rock!